Separate statistics validators and tests in different files
